v1.1 版本下发测试通过

+++ credit-business/src/main/java/com/ruoyi/business/iot/common/constant/DownCmdEnum.java
@@ -30,11 +30,11 @@ public enum DownCmdEnum {
      * udp 1.1版本命令
      */
     DOWNLINK_UDP_E0((byte)0xE0, Byte.class,1,"下发温度补偿模式"),
-    DOWNLINK_UDP_E1((byte)0xE1, Short.class,2,"补偿设定时间1"),
-    DOWNLINK_UDP_E2((byte)0xE2, Short.class,2,"补偿设定时间2"),
-    DOWNLINK_UDP_E3((byte)0xE3, Short.class,2,"补偿设定时间3"),
+    DOWNLINK_UDP_E1((byte)0xE1, Float.class,2,"补偿设定时间1"),
+    DOWNLINK_UDP_E2((byte)0xE2, Float.class,2,"补偿设定时间2"),
+    DOWNLINK_UDP_E3((byte)0xE3, Float.class,2,"补偿设定时间3"),
 
-    DOWNLINK_UDP_E4((byte)0xE4, Short.class,2,"补偿设定功率1"),
+    DOWNLINK_UDP_E4((byte)0xE4, Float.class,2,"补偿设定功率1"),
     DOWNLINK_UDP_E5((byte)0xE5, Byte.class,1,"供热模式有效性"),
